Make a wrapper object fullscreen instead of body
Making the body fullscreen doesn’t work if the canvas is not the only
element on the page. Elements like buttons, text, or images may end up
overlaid on the distorted output.

This patch injects a wrapper div into the page, moves the canvas into
it, and makes that fullscreen instead. That way the content  of the
wrapper div can be tightly controlled and no unintended elements will
be displayed.