Only listen to the fullscreen-events of one of the different vendor
Without doing this, it can happen that the handlers are called via the 'fullscreenchange' and 'webkitfullscreenchange' event when using the 'Fullscreen-API-Polyfill' (https://github.com/neovov/Fullscreen-API-Polyfill). So it is best to only listen to one of the events, otherwise the handler may be called several times which lead to nothing being presented.