Added two new modes to the polyfill
Added modes that allow the polyfill to continue working even when the
native API is present. Dev’s can choose “NO_NATIVE_API” for the
existing behavior (this is the default), “NO_NATIVE_DISPLAY” to have
the polyfill kick in when the native API is present but doesn’t provide
a display, and “ALWAYS” to always have a polyfilled VRDisplay appended
to the returned list.